+++ CHANGELOG.md
@@ -6,11 +6,13 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Nothing.
+- Command line interface for running tools (accessible through cli.php)
+- Command line utility for checking status of full text links.
 
 ### Changed
 
 - Migrated dependencies from Zend Framework to Laminas.
+- Minimum PHP version is now 7.1.3.
 - Renamed GeebyDeeby\Db\PluginManagerFactory to GeebyDeeby\ServiceManager\AbstractPluginManagerFactory for greater reusability.
 
 ### Removed